requirements.txt
postgres
django~=5.2.4
libsass
django-compressor
django-sass-processor
djangorestframework~=3.16.0